Turned a size_t into an int so that a decrement loop could work properly
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@1533 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/CSeries/byte_swapping.h
@@ -34,7 +34,7 @@ enum {
 extern void byte_swap_memory(
 	void *memory,
 	_bs_field type,
-	size_t fieldcount);
+	int fieldcount);
 
 #ifndef ALEPHONE_LITTLE_ENDIAN
 #define byte_swap_memory(memory,type,elcount) ((void)0)
